Conflicts:
	db/schema.rb

+++ db/migrate/20140318190542_remove_components.rb
@@ -1,7 +1,7 @@
 class RemoveComponents < ActiveRecord::Migration
   def up
-    drop_table :components
-    drop_table :experiment_components
+    drop_table :components if ActiveRecord::Base.connection.table_exists? 'components'
+    drop_table :experiment_components if ActiveRecord::Base.connection.table_exists? 'experiment_components'
   end
 
   def down
